skip sorting when sampling large tables
