+++ featurebyte/api/catalog.py
@@ -38,10 +38,11 @@ from featurebyte.api.target import Target
 from featurebyte.api.use_case import UseCase
 from featurebyte.api.user_defined_function import UserDefinedFunction
 from featurebyte.api.view import View
+from featurebyte.common import activate_catalog, get_active_catalog_id
 from featurebyte.common.doc_util import FBAutoDoc
 from featurebyte.exception import RecordRetrievalException
 from featurebyte.logging import get_logger
-from featurebyte.models.base import PydanticObjectId, activate_catalog, get_active_catalog_id
+from featurebyte.models.base import PydanticObjectId
 from featurebyte.models.catalog import CatalogModel
 from featurebyte.models.relationship import RelationshipType
 from featurebyte.schema.catalog import CatalogCreate, CatalogOnlineStoreUpdate, CatalogUpdate
