+++ tests/unit/migration/conftest.py
@@ -7,8 +7,8 @@ from unittest.mock import Mock
 import pytest
 from bson import ObjectId
 
+from featurebyte.common import DEFAULT_CATALOG_ID
 from featurebyte.migration.migration_data_service import SchemaMetadataService
-from featurebyte.models.base import DEFAULT_CATALOG_ID
 from featurebyte.routes.block_modification_handler import BlockModificationHandler
