+++ tests/unit/models/test_event_table.py
@@ -7,7 +7,7 @@ import datetime
 import pytest
 from pydantic import ValidationError
 
-from featurebyte.models.base import DEFAULT_CATALOG_ID
+from featurebyte.common import DEFAULT_CATALOG_ID
 from featurebyte.models.event_table import EventTableModel
 from featurebyte.models.feature_store import TableStatus
 from featurebyte.query_graph.model.feature_job_setting import FeatureJobSetting
