+++ tests/unit/query_graph/sql/adapter/test_bigquery.py
@@ -28,3 +28,23 @@ def test_random_sample(update_fixtures):
         "tests/fixtures/adapter/bigquery_random_sample.sql",
         update_fixtures,
     )
+
+
+def test_random_sample_large_table(update_fixtures):
+    """
+    Test random_sample for large table
+    """
+    output = sql_to_string(
+        BigQueryAdapter.random_sample(
+            select("a", "b").from_("table1"),
+            desired_row_count=100,
+            total_row_count=10000001,
+            seed=1234,
+        ),
+        SourceType.BIGQUERY,
+    )
+    assert_equal_with_expected_fixture(
+        output,
+        "tests/fixtures/adapter/bigquery_random_sample_large_table.sql",
+        update_fixtures,
+    )
